+++ hata/discord/guild/audit_logs/preinstanced.py
@@ -8,6 +8,7 @@ from ...bases import Preinstance as P, PreinstancedBase
 
 from .change_converters.all_ import MERGED_CONVERTERS
 from .change_converters.application_command import APPLICATION_COMMAND_CONVERTERS
+from .change_converters.auto_moderation import AUTO_MODERATION_CONVERTERS
 from .change_converters.channel import CHANNEL_CONVERTERS
 from .change_converters.channel_permission_overwrite import CHANNEL_PERMISSION_OVERWRITE_CONVERTERS
 from .change_converters.emoji import EMOJI_CONVERTERS
@@ -21,10 +22,10 @@ from .change_converters.sticker import STICKER_CONVERTERS
 from .change_converters.user import USER_CONVERTERS
 from .change_converters.webhook import WEBHOOK_CONVERTERS
 from .target_converters import (
-    target_converter_application_command, target_converter_channel, target_converter_emoji, target_converter_guild,
-    target_converter_integration, target_converter_invite, target_converter_none, target_converter_role,
-    target_converter_scheduled_event, target_converter_stage, target_converter_sticker, target_converter_thread,
-    target_converter_user, target_converter_webhook
+    target_converter_application_command, target_converter_auto_moderation, target_converter_channel,
+    target_converter_emoji, target_converter_guild, target_converter_integration, target_converter_invite,
+    target_converter_none, target_converter_role, target_converter_scheduled_event, target_converter_stage,
+    target_converter_sticker, target_converter_thread, target_converter_user, target_converter_webhook
 )
 
 
@@ -89,6 +90,8 @@ class AuditLogTargetType(PreinstancedBase):
     +-------------------------------+-------------------------------+-------+-------------------------------------------+
     | thread                        | thread                        | 14    | CHANNEL_CONVERTERS                        |
     +-------------------------------+-------------------------------+-------+-------------------------------------------+
+    | auto_moderation               | auto moderation               | 15    | AUTO_MODERATION_CONVERTERS                |
+    +-------------------------------+-------------------------------+-------+-------------------------------------------+
     """
     INSTANCES = {}
     VALUE_TYPE = int
@@ -123,6 +126,7 @@ class AuditLogTargetType(PreinstancedBase):
         APPLICATION_COMMAND_CONVERTERS,
     )
     thread = P(14, 'thread', target_converter_thread, CHANNEL_CONVERTERS)
+    auto_moderation = P(15, 'auto moderation', target_converter_auto_moderation ,AUTO_MODERATION_CONVERTERS)
     
     
     @copy_docs(PreinstancedBase.__repr__)
@@ -293,6 +297,16 @@ class AuditLogEvent(PreinstancedBase):
     +---------------------------------------+---------------------------------------+-------+-------------------------------+
     | thread_delete                         | thread delete                         | 112   | thread                        |
     +---------------------------------------+---------------------------------------+-------+-------------------------------+
+    | application_command_permission_update | application command permission update | 121   | none                          |
+    +---------------------------------------+---------------------------------------+-------+-------------------------------+
+    | auto_moderation_rule_create           | auto moderation rule create           | 140   | auto_moderation               |
+    +---------------------------------------+---------------------------------------+-------+-------------------------------+
+    | auto_moderation_rule_update           | auto moderation rule update           | 141   | auto_moderation               |
+    +---------------------------------------+---------------------------------------+-------+-------------------------------+
+    | auto_moderation_rule_delete           | auto moderation rule delete           | 142   | auto_moderation               |
+    +---------------------------------------+---------------------------------------+-------+-------------------------------+
+    | auto_moderation_block_message         | auto moderation block message         | 143   | auto_moderation               |
+    +---------------------------------------+---------------------------------------+-------+-------------------------------+
     """
     INSTANCES = {}
     VALUE_TYPE = int
@@ -379,6 +393,11 @@ class AuditLogEvent(PreinstancedBase):
         'application command permission update',
         AuditLogTargetType.application_command,
     )
+    
+    auto_moderation_rule_create = P(140, 'auto moderation rule create', AuditLogTargetType.auto_moderation)
+    auto_moderation_rule_update = P(141, 'auto moderation rule update', AuditLogTargetType.auto_moderation)
+    auto_moderation_rule_delete = P(142, 'auto moderation rule delete', AuditLogTargetType.auto_moderation)
+    auto_moderation_block_message = P(143, 'auto moderation block message', AuditLogTargetType.auto_moderation)
     
     
     def __init__(self, value, name, target_type):