+++ hata/discord/guild/audit_logs/target_converters.py
@@ -123,3 +123,7 @@ def target_converter_thread(entry):
 
 def target_converter_application_command(entry):
     return None
+
+
+def target_converter_auto_moderation(entry):
+    return None